+++ examples/hybrid_rrf/model.rb
@@ -12,16 +12,16 @@ require_relative "../../lib/parade_db"
 class MockItem < ActiveRecord::Base
   include ParadeDB::Model
 
-  self.table_name = "mock_items_hybrid_rrf"
+  self.table_name = "mock_items"
   self.primary_key = "id"
 
   has_neighbors :embedding
 end
 
 class MockItemIndex < ParadeDB::Index
-  self.table_name = :mock_items_hybrid_rrf
+  self.table_name = :mock_items
   self.key_field = :id
-  self.index_name = :mock_items_hybrid_rrf_bm25_idx
+  self.index_name = :search_idx
   self.fields = {
     id: nil,
     description: nil,